+++ scripts/lc-builds/blueos_pgi.sh
@@ -7,7 +7,20 @@
 # SPDX-License-Identifier: (BSD-3-Clause)
 ###############################################################################
 
-BUILD_SUFFIX=lc_blueos-gcc-8.3.1
+if [ "$1" == "" ]; then
+  echo
+  echo "You must pass a compiler version number to script. For example,"
+  echo "    blueos_pgi.sh 21.1"
+  exit
+fi
+
+COMP_VER=$1
+
+BUILD_SUFFIX=lc_blueos-pgi-${COMP_VER}
+
+echo
+echo "Creating build directory ${BUILD_SUFFIX} and generating configuration in it"
+echo
 
 rm -rf build_${BUILD_SUFFIX} 2>/dev/null
 mkdir build_${BUILD_SUFFIX} && cd build_${BUILD_SUFFIX}
@@ -16,8 +29,9 @@ module load cmake/3.14.5
 
 cmake \
   -DCMAKE_BUILD_TYPE=Release \
-  -DCMAKE_CXX_COMPILER=/usr/tce/packages/gcc/gcc-8.3.1/bin/g++ \
-  -C ../host-configs/lc-builds/blueos/gcc_X.cmake \
+  -DCMAKE_CXX_COMPILER=/usr/tce/packages/pgi/pgi-${COMP_VER}/bin/pgc++ \
+  -DCMAKE_C_COMPILER=/usr/tce/packages/pgi/pgi-${COMP_VER}/bin/pgcc \
+  -C ../host-configs/lc-builds/blueos/pgi_X.cmake \
   -DENABLE_OPENMP=On \
   -DCMAKE_INSTALL_PREFIX=../install_${BUILD_SUFFIX} \
   "$@" \
